+++ src/Game/Werewolf/Messages.hs
@@ -43,9 +43,6 @@ module Game.Werewolf.Messages (
     -- * Angel's turn messages
     angelJoinedVillagersMessage,
 
-    -- * Bear Tamer's turn messages
-    ursusGruntsMessage,
-
     -- * Defender's turn messages
 
     -- ** Error messages
@@ -53,6 +50,9 @@ module Game.Werewolf.Messages (
 
     -- * Devoted Servant's turn messages
     devotedServantRevealedMessage, devotedServantJoinedPackMessages,
+
+    -- * Druid's turn messages
+    ferinaGruntsMessage,
 
     -- * Scapegoat's turn messages
     scapegoatChoseAllowedVotersMessage,
@@ -148,13 +148,13 @@ stageMessages :: Game -> [Message]
 stageMessages game = case game ^. stage of
     DefendersTurn       -> defendersTurnMessages defendersName
     DevotedServantsTurn -> devotedServantsTurnMessages devotedServantsName victimsName
+    FerinasGrunt        -> []
     GameOver            -> []
     Lynching            -> []
     ScapegoatsTurn      -> scapegoatsTurnMessages scapegoatsName
     SeersTurn           -> seersTurnMessages seersName
     Sunrise             -> [sunriseMessage]
     Sunset              -> [nightFallsMessage]
-    UrsussGrunt         -> []
     VillagesTurn        -> if isFirstRound game
         then firstVillagesTurnMessages
         else villagesTurnMessages
@@ -394,11 +394,11 @@ pingRoleMessage :: Text -> Message
 pingRoleMessage roleName = publicMessage $ T.concat ["Waiting on the ", roleName, "..."]
 
 currentStageMessages :: Text -> Stage -> [Message]
+currentStageMessages _ FerinasGrunt = []
 currentStageMessages to GameOver    = [gameIsOverMessage to]
 currentStageMessages _ Lynching     = []
 currentStageMessages _ Sunrise      = []
 currentStageMessages _ Sunset       = []
-currentStageMessages _ UrsussGrunt  = []
 currentStageMessages to turn        = [privateMessage to $ T.concat
     [ "It's currently the ", showTurn turn, " turn."
     ]]
@@ -406,13 +406,13 @@ currentStageMessages to turn        = [privateMessage to $ T.concat
         showTurn :: Stage -> Text
         showTurn DefendersTurn          = "Defender's"
         showTurn DevotedServantsTurn    = "Devoted Servant's"
+        showTurn FerinasGrunt           = undefined
         showTurn GameOver               = undefined
         showTurn Lynching               = undefined
         showTurn ScapegoatsTurn         = "Scapegoat's"
         showTurn SeersTurn              = "Seer's"
         showTurn Sunrise                = undefined
         showTurn Sunset                 = undefined
-        showTurn UrsussGrunt            = undefined
         showTurn VillagesTurn           = "village's"
         showTurn WerewolvesTurn         = "Werewolves'"
         showTurn WildChildsTurn         = "Wild-child's"
@@ -458,12 +458,6 @@ angelJoinedVillagersMessage = publicMessage $ T.unwords
     , "Now he is stuck here, doomed forever to live out a mortal life as a Villager."
     ]
 
-ursusGruntsMessage :: Message
-ursusGruntsMessage = publicMessage $ T.unwords
-    [ "Ursus wakes from his slumber, disturbed and on edge."
-    , "He loudly grunts as he smells danger."
-    ]
-
 playerCannotProtectSamePlayerTwiceInARowMessage :: Text -> Message
 playerCannotProtectSamePlayerTwiceInARowMessage to =
     privateMessage to "You cannot protect the same player twice in a row!"
@@ -492,6 +486,12 @@ devotedServantJoinedPackMessages devotedServantsName werewolfNames =
                 [ "As you enter you see his pack", T.intercalate ", " werewolfNames
                 , "waiting for you."
                 ]]
+
+ferinaGruntsMessage :: Message
+ferinaGruntsMessage = publicMessage $ T.unwords
+    [ "Ferina wakes from her slumber, disturbed and on edge."
+    , "She loudly grunts as she smells danger."
+    ]
 
 scapegoatChoseAllowedVotersMessage :: [Text] -> Message
 scapegoatChoseAllowedVotersMessage allowedVoters = publicMessage $ T.unwords