+++ CHANGELOG.md
@@ -2,10 +2,15 @@
 
 ### Upcoming
 
+*Minor*
+
+* Renamed the Bear Tamer to Druid. ([#131](https://github.com/hjwylde/werewolf/issues/131))
+
 *Revisions*
 
 * Updated the Witch's description and rules. ([#138](https://github.com/hjwylde/werewolf/issues/138))
 * Updated the Scapegoat's description. ([#133](https://github.com/hjwylde/werewolf/issues/133))
+* Updated the Druid's description and rules. ([#131](https://github.com/hjwylde/werewolf/issues/131))
 
 ### v0.4.11.0